+++ source/extensions/omni.isaac.range_sensor/docs/CHANGELOG.md
@@ -2,6 +2,16 @@
 CHANGELOG
 **********
 
+
+
+[0.2.1] - 2022-03-04
+========================
+
+Changed
+========
+- Lidar sensor is now fully multithreaded per sensor and per ray
+- Improved use of tasking framework and simplified scan logic
+
 [0.2.0] - 2022-02-11
 ========================